Type change in header

+++ src/common/utils_flat.h
@@ -3,7 +3,7 @@ int sqr(int x);
 
 double *empty_matrix(int matrix_dimention);
 void delete_matrix(double *matrix);
-double *generate_matrix(int matrix_dimention, int max_value);
+double *generate_matrix(int matrix_dimention, double max_value);
 
 enum matrix_prefil
 {
